hook up new workers
