+++ package.json
@@ -134,6 +134,7 @@
   },
   "workspaces": [
     "jacdac-ts",
+    "jacscript/vm",
     "jacscript/compiler"
   ],
   "scripts": {
@@ -160,16 +161,16 @@
     "buildworkercsv": "cd src/workers/csv && node ../../../node_modules/microbundle/dist/cli.js build -f modern --no-compress --types --ts-config tsconfig.json",
     "buildworkerdata": "cd src/workers/data && node ../../../node_modules/microbundle/dist/cli.js build -f modern --no-compress --types --ts-config tsconfig.json",
     "buildworkertf": "cd src/workers/tf && node ../../../node_modules/microbundle/dist/cli.js build -f modern --no-compres --types --ts-config tsconfig.json",
-    "buildworkervm": "cd src/workers/vm && node ../../../node_modules/microbundle/dist/cli.js build -f modern --no-compress --types --ts-config tsconfig.json",
-    "buildworkerjacscript": "cd src/workers/jacscript && node ../../../node_modules/esbuild/bin/esbuild --bundle jacscript.worker.ts --outdir=dist/node_modules --sourcemap",
+    "buildworkervm": "cd src/workers/vm && node ../../../node_modules/esbuild/bin/esbuild --bundle vm-worker.ts --outdir=dist/node_modules --sourcemap",
+    "buildworkerjacscript": "cd src/workers/jacscript && node ../../../node_modules/esbuild/bin/esbuild --bundle jacscript-worker.ts --outdir=dist/node_modules --sourcemap",
     "buildworkercad": "cd src/workers/cad && node ../../../node_modules/microbundle/dist/cli.js build -f modern --no-compress --types --ts-config tsconfig.json",
     "buildworkers": "yarn buildworkercsv && yarn buildworkerdata && yarn buildworkertf && yarn buildworkervm && yarn buildworkercad && yarn buildworkerjacscript",
     "watchdocsts": "node node_modules/typescript/bin/tsc -b src --watch",
     "watchworkercsv": "cd src/workers/csv && node ../../../node_modules/microbundle/dist/cli.js watch -f modern --no-compress --types --ts-config tsconfig.json",
     "watchworkerdata": "cd src/workers/data && node ../../../node_modules/microbundle/dist/cli.js watch -f modern --no-compress --types --ts-config tsconfig.json",
     "watchworkertf": "cd src/workers/tf && node ../../../node_modules/microbundle/dist/cli.js watch -f modern --no-compress --types --ts-config tsconfig.json",
-    "watchworkervm": "cd src/workers/vm && node ../../../node_modules/microbundle/dist/cli.js watch -f modern --no-compress --types --ts-config tsconfig.json",
-    "watchworkerjacscript": "cd src/workers/jacscript && node ../../../node_modules/esbuild/bin/esbuild --bundle jacscript.worker.ts --outdir=dist/node_modules --sourcemap --watch",
+    "watchworkervm": "cd src/workers/vm && node ../../../node_modules/esbuild/bin/esbuild --bundle vm-worker.ts --outdir=dist/node_modules --sourcemap --watch",
+    "watchworkerjacscript": "cd src/workers/jacscript && node ../../../node_modules/esbuild/bin/esbuild --bundle jacscript-worker.ts --outdir=dist/node_modules --sourcemap --watch",
     "watchworkercad": "cd src/workers/cad && node ../../../node_modules/microbundle/dist/cli.js watch -f modern --no-compress --types --ts-config tsconfig.json",
     "watchts": "cd jacdac-ts && yarn watch",
     "watchjacscript": "cd jacscript/compiler && yarn watch",
