+++ src/components/blockly/dsl/workers/vm.proxy.ts
@@ -1,20 +1,15 @@
 import type {
     VMStateResponse,
     VMState,
-    VMCommandRequest,
     VMRequest,
     VMPacketRequest,
-    VMDeployRequest,
-} from "../../../../workers/vm/dist/node_modules/vm.worker"
+} from "../../../../workers/vm/dist/node_modules/vm-worker"
 import workerProxy, { WorkerProxy } from "./proxy"
-import bus from "../../../../jacdac/providerbus"
 import {
     CHANGE,
     MESSAGE,
-    REPORT_UPDATE,
 } from "../../../../../jacdac-ts/src/jdom/constants"
 import { JDBridge } from "../../../../../jacdac-ts/src/jdom/bridge"
-import { JacscriptManagerServer } from "../../../../../jacdac-ts/src/servers/jacscriptmanagerserver"
 
 class JacscriptBridge extends JDBridge {
     state: VMState = "stopped"
@@ -53,87 +48,6 @@ class JacscriptBridge extends JDBridge {
     }
 }
 
-async function jacscriptDeploy(
-    binary: Uint8Array,
-    debugInfo: unknown,
-    restart?: boolean
-): Promise<VMStateResponse> {
-    const bridge = jacscriptBridge()
-    console.log(`jdvm proxy: deploy ${binary.length} bytes`)
-    const res = await bridge.worker.postMessage<
-        VMDeployRequest,
-        VMStateResponse
-    >({
-        worker: "vm",
-        type: "deploy",
-        binary,
-        debugInfo,
-        restart,
-    })
-    return res
-}
-
-async function jacscriptCommand(
-    action: "start" | "stop"
-): Promise<VMStateResponse> {
-    const bridge = jacscriptBridge()
-    if (action === "start") bridge.bus = bus
-    else bridge.bus = undefined
-    console.log(`jdvm: command ${action}`)
-    const res = await bridge.worker.postMessage<
-        VMCommandRequest,
-        VMStateResponse
-    >({
-        worker: "vm",
-        type: "command",
-        action,
-    })
-    return res
-}
-
-class VMJacscriptManagerServer extends JacscriptManagerServer {
-    bridge: JacscriptBridge
-    state: VMState
-
-    constructor() {
-        super()
-
-        this.bridge = jacscriptBridge()
-        this.bridge.on(CHANGE, this.handleBridgeChange.bind(this))
-        this.running.on(CHANGE, this.handleRunningChange.bind(this))
-
-        this.on(
-            JacscriptManagerServer.PROGRAM_CHANGE,
-            this.handleProgramChange.bind(this)
-        )
-    }
-
-    private handleProgramChange() {
-        console.debug("jdvm proxy: program change")
-        const autoStart = !!this.autoStart.values()[0]
-        jacscriptDeploy(this.binary, this.debugInfo, !!autoStart)
-    }
-
-    private handleBridgeChange() {
-        this.state = this.bridge.state
-        const running =
-            this.state === "running" || this.state === "initializing"
-        console.log(
-            `jdvm server: state ${this.state} ${
-                running ? "running" : "stopped"
-            }`
-        )
-        this.running.setValues([running], true)
-    }
-
-    private handleRunningChange() {
-        const running = this.running.values()[0]
-        const action = running ? "start" : "stop"
-        console.log(`jdvm server: ${action} ${this.state}`)
-        jacscriptCommand(action)
-    }
-}
-
 let bridge: JacscriptBridge
 export function jacscriptBridge() {
     if (!bridge) {
@@ -141,9 +55,4 @@ export function jacscriptBridge() {
         bridge = new JacscriptBridge(worker)
     }
     return bridge
-}
-let server: JacscriptManagerServer
-export function createVMJacscriptManagerServer(): JacscriptManagerServer {
-    if (!server) server = new VMJacscriptManagerServer()
-    return server
 }
